Create Pirple_PythonIsEasy_Connect4.py
When I run this code (and it is not in its finished form, but I can't get it to do something simple), the row/column allocation when I ta a turn does not make sense. Multiple spaces are populated sometimes, and other times nothing happens. But when I try to populate a space twice, it returns that the space is occupied even though it does not show up in the grid.