Create ClassesHW9.py
Homework 9 from "Python is Easy"...unfortunately they kicked me out of the student forum because I called them out on not inspecting the homework close enough.  The assignment is as follows:

Details:
 
Create a class called "Vehicle" and methods that allow you to set the "Make", "Model", "Year,", and "Weight". (done)

The class should also contain a "NeedsMaintenance" boolean that defaults to False, and and "TripsSinceMaintenance" Integer that defaults to 0. (not sure I have this right. These two methods are not being recognized in the code (unresolved))

Next an inheritance classes from Vehicle called "Cars". (Done) 

The Cars class should contain a method called "Drive" that sets the state of a boolean isDriving to True.  It should have another method called "Stop" that sets the value of isDriving to false. (definitely having problems understanding this)

Switching isDriving from true to false should increment the "TripsSinceMaintenance" counter. And when TripsSinceMaintenance exceeds 100, then the NeedsMaintenance boolean should be set to true. (way over my head)

Add a "Repair" method to either class that resets the TripsSinceMaintenance to zero, and NeedsMaintenance to false. (haven't even touched it yet)

Create 3 different cars, using your Cars class, and drive them all a different number of times. Then print out their values for Make, Model, Year, Weight, NeedsMaintenance, and TripsSinceMaintenance

Extra Credit:

Create a Planes class that is also an inheritance class from Vehicle. Add methods to the Planes class for Flying and Landing (similar to Driving and Stopping), but different in one respect: Once the NeedsMaintenance boolean gets set to true, any attempt at flight should be rejected (return false), and an error message should be printed saying that the plane can't fly until it's repaired.